requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
langchain-ollama
langchain-huggingface
langchain-qdrant
qdrant-client
sentence-transformers
pydantic
python-dotenv
